requirements.txt
chainlit==1.0.0
langchain==0.2.4
langchain_community
langchain_openai
langchain_core==0.2.40
langchain_huggingface
python-dotenv
tiktoken
pymupdf
qdrant-client
sentence-transformers
huggingface-hub==0.24.6
